requirements.txt
asgiref==3.4.1
CacheControl==0.12.10
cachy==0.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
cleo==0.8.1
clikit==0.6.2
cloudinary==1.28.0
colorama==0.4.4
crashtest==0.3.1
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.9
django-heroku==0.3.1
django-phonenumber-field==6.0.0
django-uuidfield==0.5.0
djangorestframework==3.12.4
filelock==3.5.0
gunicorn==20.1.0
html5lib==1.1
idna==3.3
image==1.5.33
importlib-metadata==4.11.1
keyring==23.5.0
lockfile==0.12.2
msgpack==1.0.3
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
phonenumberslite==8.12.37
Pillow==9.0.1
pipenv==2022.1.8
pkginfo==1.8.2
platformdirs==2.5.0
poetry==1.1.13
poetry-core==1.0.7
psycopg2==2.9.2
ptyprocess==0.7.0
pygame==2.1.2
pylev==1.4.0
pyparsing==3.0.7
python-decouple==3.5
python-dotenv==0.19.2
pytz==2021.3
pywin32-ctypes==0.2.0
qrcode==7.3.1
requests==2.27.1
requests-toolbelt==0.9.1
shellingham==1.4.0
six==1.16.0
sqlparse==0.4.2
tomlkit==0.9.2
tzdata==2021.5
urllib3==1.26.7
virtualenv==20.13.1
virtualenv-clone==0.5.7
webencodings==0.5.1
whitenoise==5.3.0
zipp==3.7.0
